LLM_Weather/chatbot/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.3.2
python-dotenv==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
pydantic==2.5.0
aiohttp==3.9.1
pydantic-settings==2.1.0
